Move command into core commands
